Started some (tiny amount of) work on auth/passport. Server-side: Added a method to extract Bearer-token (and a test for this method)
